update ssn into uint32

+++ contracts/books/boa/IBookOfIA.sol
@@ -80,7 +80,7 @@ interface IBookOfIA {
 
     function isRegistered(address body) external view returns (bool);
 
-    function counterOfDocs() external view returns (uint16);
+    function counterOfDocs() external view returns (uint32);
 
     function passedReview(address body) external view returns (bool);
